完善角色管理功能

+++ src/main/java/com/roncoo/adminlte/biz/RoleBiz.java
@@ -1,12 +1,23 @@
 package com.roncoo.adminlte.biz;
 
+import java.util.ArrayList;
+import java.util.HashSet;
+import java.util.List;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Transactional;
 
 import com.roncoo.adminlte.bean.Result;
+import com.roncoo.adminlte.bean.entity.RcPermission;
 import com.roncoo.adminlte.bean.entity.RcRole;
+import com.roncoo.adminlte.bean.entity.RcRolePermissions;
+import com.roncoo.adminlte.bean.vo.RcRoleVo;
+import com.roncoo.adminlte.service.PermissionService;
+import com.roncoo.adminlte.service.RolePermissionsService;
 import com.roncoo.adminlte.service.RoleService;
 import com.roncoo.adminlte.util.base.Page;
+import com.roncoo.adminlte.util.base.StringUtils;
 
 @Component
 public class RoleBiz {
@@ -14,20 +25,77 @@ public class RoleBiz {
 	@Autowired
 	private RoleService service;
 
-	public Result<RcRole> query(long id) {
-		return service.query(id);
+	@Autowired
+	private RolePermissionsService rolePermissionsService;
+
+	@Autowired
+	private PermissionService permissionService;
+
+	public Result<RcRoleVo> query(long id) {
+		Result<RcRole> result = service.query(id);
+		Result<RcRoleVo> resultRoleVo = new Result<RcRoleVo>();
+		if (result.isStatus()) {
+			RcRoleVo rcRoleVo = new RcRoleVo(result.getResultData());
+			HashSet<String> permissionNameSet = new HashSet<String>();
+			HashSet<String> permissionSet = new HashSet<String>();
+			Result<List<RcRolePermissions>> resultRolePermissions = rolePermissionsService.queryByRoleId(id);
+			if (resultRolePermissions.isStatus()) {
+				ArrayList<Long> idList = new ArrayList<Long>();
+				for (RcRolePermissions rcRolePermissions : resultRolePermissions.getResultData()) {
+					idList.add(rcRolePermissions.getPermissionId());
+				}
+				Result<List<RcPermission>> resultPermission = permissionService.listForId(idList);
+				for (RcPermission rcPermission : resultPermission.getResultData()) {
+					permissionNameSet.add(rcPermission.getPermissionsName());
+					permissionSet.add(rcPermission.getPermissionsValue());
+				}
+				rcRoleVo.setPermission(StringUtils.toString(permissionSet));
+				rcRoleVo.setPermissionName(StringUtils.toString(permissionNameSet));
+				rcRoleVo.setPermissions(permissionSet);
+				rcRoleVo.setPermissionNames(permissionNameSet);
+
+				resultRoleVo.setErrCode(0);
+				resultRoleVo.setStatus(true);
+				resultRoleVo.setResultData(rcRoleVo);
+				resultRoleVo.setErrMsg("查询成功");
+				return resultRoleVo;
+			}
+		}
+		return resultRoleVo;
 	}
 
-	public Result<Integer> save(RcRole rcRole) {
-		return service.save(rcRole);
+	public Result<List<RcPermission>> queryPermissionList() {
+		return permissionService.list();
 	}
 
-	public Result<Integer> update(RcRole rcRole) {
-		return service.update(rcRole);
+	@Transactional
+	public Result<Integer> save(RcRole rcRole, List<Long> permissionList) {
+		Result<Integer> result = service.save(rcRole);
+		if (result.isStatus()) {
+			Result<RcRole> resultRole = service.queryByRoleName(rcRole.getRoleName());
+			if (result.isStatus()) {
+				return rolePermissionsService.save(resultRole.getResultData().getId(), permissionList);
+			}
+		}
+		return result;
 	}
 
+	@Transactional
+	public Result<Integer> update(RcRole rcRole, List<Long> permissionList) {
+		Result<Integer> result = rolePermissionsService.update(rcRole.getId(), permissionList);
+		if (result.isStatus()) {
+			return service.update(rcRole);
+		}
+		return result;
+	}
+
+	@Transactional
 	public Result<Integer> delete(long id) {
-		return service.delete(id);
+		Result<Integer> result = rolePermissionsService.delete(id);
+		if (result.isStatus()) {
+			return service.delete(id);
+		}
+		return result;
 	}
 
 	public Result<Page<RcRole>> listForPage(int pageCurrent, int pageSize, String date, String search) {
